UP-629 Anchoring V2 #comment Adding a couple of important entities to better control the new procedure.
