+++ event-log-chainer/src/main/scala/com/ubirch/chainer/services/TreeCache.scala
@@ -2,10 +2,10 @@ package com.ubirch.chainer.services
 
 import com.typesafe.config.Config
 import com.ubirch.chainer.models.Mode
-import com.ubirch.models.{Cache, MemCache}
+import com.ubirch.models.{ Cache, MemCache }
 import javax.inject._
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.{ ExecutionContext, Future }
 
 class TreeCache @Inject() (@Named(MemCache.name) cache: Cache, config: Config)(implicit ec: ExecutionContext) {
 
@@ -13,11 +13,11 @@ class TreeCache @Inject() (@Named(MemCache.name) cache: Cache, config: Config)(i
 
   val mode: Mode = Mode.getMode(modeFromConfig)
 
-  val LATEST_TREE: String = Mode.fold(mode)("LASTEST_SLAVE_TREE")( "LATEST_MASTER_TREE")
+  val LATEST_TREE: String = Mode.fold(mode)("LASTEST_SLAVE_TREE")("LATEST_MASTER_TREE")
 
   def prefix(value: String): String = {
     val px = Mode.fold(mode)("sl.")(("ml."))
-    if(!value.startsWith(px)) px + value else value
+    if (!value.startsWith(px)) px + value else value
   }
 
   def latest: Future[Option[String]] = cache.get(LATEST_TREE)
